# Documentation/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant and owner created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponseDto" }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "summary": "Register a new tenant and owner user",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponseDto" }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "summary": "Authenticate user and return tokens",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access token issued",
            "content": {
              "application/json": {
                "schema": {
                  "properties": { "accessToken": { "type": "string" } }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "summary": "Exchange a refresh token for a new access token",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenDto" }
            }
          }
        },
        "responses": { "200": { "description": "Token revoked" } },
        "summary": "Revoke a refresh token",
        "tags": ["Auth"]
      }
    },
...
  },
  "info": {
    "title": "Persona Finance API",
    "description": "Multi-tenant finance system. Standard error response format includes: `statusCode`, `message`, `errors[]`, `timestamp`, `path`, `requestId`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "example": "Acme Trading Co.",
            "description": "Name of the business / tenant"
          },
          "fullName": {
            "type": "string",
            "example": "Zaeem Hassan",
            "description": "Owner full name"
          },
          "email": {
            "type": "string",
            "example": "zaeem@acme.com",
            "description": "Owner email address"
          },
          "password": {
            "type": "string",
            "example": "MyPass123",
            "description": "Min 8 chars, 1 uppercase, 1 lowercase, 1 number"
          }
        },
        "required": ["tenantName", "fullName", "email", "password"]
      },
      "TenantSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "6c6f7f48-3d5b-4a3f-9b1d-9c0d73b0c3d2"
          },
          "name": { "type": "string", "example": "Acme Trading Co." },
          "baseCurrency": { "type": "string", "example": "PKR" },
          "timezone": { "type": "string", "example": "Asia/Karachi" }
        },
        "required": ["id", "name", "baseCurrency", "timezone"]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "d2f2c7b5-0c2a-4aa2-9c60-6b3f94b7e8d4"
          },
          "tenantId": {
            "type": "string",
            "example": "6c6f7f48-3d5b-4a3f-9b1d-9c0d73b0c3d2"
          },
          "fullName": { "type": "string", "example": "Zaeem Hassan" },
          "email": { "type": "string", "example": "zaeem@acme.com" },
          "role": { "type": "string", "example": "OWNER" },
          "tenant": { "$ref": "#/components/schemas/TenantSummaryDto" }
        },
        "required": ["id", "tenantId", "fullName", "email", "role"]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": { "$ref": "#/components/schemas/AuthUserDto" }
        },
        "required": ["accessToken", "refreshToken", "user"]
      },
      "ApiErrorField": {
        "type": "object",
        "properties": {
          "field": { "type": "string", "example": "name" },
          "message": { "type": "string", "example": "must not be empty" }
        },
        "required": ["field", "message"]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "message": { "type": "string", "example": "Validation failed" },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ApiErrorField" }
          },
          "timestamp": {
            "type": "string",
            "example": "2026-02-11T12:00:00.000Z"
          },
          "path": { "type": "string", "example": "/api/v1/customers" },
          "requestId": { "type": "string", "example": "req_123" }
        },
        "required": ["statusCode", "message", "timestamp", "path"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "zaeem@acme.com" },
          "password": { "type": "string", "example": "MyPass123" }
        },
        "required": ["email", "password"]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": ["refreshToken"]
      },
...
    }
  }
}